feat: move types and interfaces to shared
